requirements.txt
numpy
Pillow
requests
python-dotenv
